initially completed add to outfit button functionality
