+++ client/src/components/RelatedItems.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import RelatedProductsList from './related-items/RelatedProductsList.jsx';
 import YourOutfitList from './related-items/YourOutfitList.jsx';
 
-const RelatedItems = ({currentRelatedProducts, getAvgRating, currentProduct}) => {
+const RelatedItems = ({currentRelatedProducts, getAvgRating, currentProduct, currentProductStyles, currentProductAvgRating}) => {
 
 
   return (
@@ -14,7 +14,7 @@ const RelatedItems = ({currentRelatedProducts, getAvgRating, currentProduct}) =>
 
     <hr></hr>
     <div><h3>Your Outfit </h3> </div>
-    <YourOutfitList getAvgRating = {getAvgRating} currentProduct = {currentProduct} />
+    <YourOutfitList getAvgRating = {getAvgRating} currentProduct = {currentProduct} currentProductStyles = {currentProductStyles} currentProductAvgRating = {currentProductAvgRating} />
 
     </>
