+++ client/src/components/related-items/RelatedCard.jsx
@@ -13,8 +13,8 @@ const RelatedCard = ({index, product, id, name, category, price, avgRating, feat
           <img
             src={image}
             width="175"
-            height="175"
-            alt="header image"
+            height="200"
+            alt="product image"
           /> <br></br>
           <div className = 'card-details'>
           <small className = 'card-text'>{category} </small>
